Added a gameobject that tracks data about the player from room to room
It only tracks health and ammo. The gameobject is designed to be a Singleton, meaning there should always only be one instance of it. It is instantiated in the main menu. If you run the game from the main menu, your data will carry from scene to scene. Otherwise, it will not.